+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.5.5] - 2025-11-20
+
+### Added
+- Additional fallback to CDN if Microsoft 365 apps fail to install via Winget.
+- Added additional error messages during application installation to keep users informed of the process status.
+
+---
+
 ## [0.5.4] - 2025-11-19
 
 ### Added
@@ -102,6 +110,7 @@ First open-source release of WinDeploy - Windows Deployment Automation Toolkit. 
 
 ---
 
+[0.5.5]: https://github.com/Stensel8/WinDeploy/releases/tag/v0.5.5
 [0.5.4]: https://github.com/Stensel8/WinDeploy/releases/tag/v0.5.4
 [0.5.3]: https://github.com/Stensel8/WinDeploy/releases/tag/v0.5.3
 [0.5.2]: https://github.com/Stensel8/WinDeploy/releases/tag/v0.5.2